TG_BOT/requirements.txt
geographiclib==2.0
geopy==2.3.0
google==3.0.0
googlesearch-python==1.2.3
python-telegram-bot==13.7
requests==2.31.0
urllib3==2.0.2
wikipedia==1.4.0